refactor: restructure tft-bridge pallet #735
